refactor(gateway): GAR-440 slice 10.a — extract bootstrap/config.rs
TDD REFACTOR puro (zero behavior change). Slice 10.a do epic
GAR-430 / Q10 (Quality Gates Phase 3.6). Primeira extração do
refactor de bootstrap.rs (2415 LOC) em 6 PRs pequenos.

Movido para o novo `bootstrap/config.rs`:
- pub(crate) fn default_vault_path()
- pub(super) fn default_allowlist_path()  (era fn privada)
- pub(crate) fn resolve_api_key()
- 3 unit tests resolve_api_key_*

Path layout: bootstrap.rs renomeado para bootstrap/mod.rs (git mv).
Re-exports em mod.rs preservam crate::bootstrap::default_vault_path
e crate::bootstrap::resolve_api_key, consumidos por admin/handlers,
router e state.

Validação local:
- cargo fmt --check ✓
- cargo check -p garraia-gateway --locked ✓
- cargo +1.92 check --workspace --exclude garraia-desktop --locked ✓
- cargo clippy -p garraia-gateway --all-targets -- -D warnings ✓
- cargo test -p garraia-gateway --lib bootstrap:: → 5/5 passed

LOC bootstrap/mod.rs: 2415 → 2353 (62 removidas)
LOC bootstrap/config.rs: novo, 79 linhas (3 fns + 3 tests + docs)
