+++ crates/garraia-gateway/src/bootstrap/mod.rs
@@ -22,42 +22,14 @@ use tracing::{error, info, warn};
 
 use crate::state::SharedState;
 
-/// Default vault path under the user's home directory.
-pub(crate) fn default_vault_path() -> Option<PathBuf> {
-    Some(
-        garraia_config::ConfigLoader::default_config_dir()
-            .join("credentials")
-            .join("vault.json"),
-    )
-}
+mod config;
 
-fn default_allowlist_path() -> PathBuf {
-    garraia_config::ConfigLoader::default_config_dir().join("allowlist.json")
-}
-
-/// Resolve an API key using the priority chain: vault -> config -> env var.
-pub(crate) fn resolve_api_key(
-    config_key: Option<&str>,
-    vault_credential_key: &str,
-    env_var: &str,
-) -> Option<String> {
-    // 1. Try credential vault (only works when GARRAIA_VAULT_PASSPHRASE is set)
-    if let Some(vault_path) = default_vault_path()
-        && let Some(val) = garraia_security::try_vault_get(&vault_path, vault_credential_key)
-    {
-        return Some(val);
-    }
-
-    // 2. Config file value
-    if let Some(key) = config_key
-        && !key.is_empty()
-    {
-        return Some(key.to_string());
-    }
-
-    // 3. Environment variable
-    std::env::var(env_var).ok()
-}
+// Slice 10.a (GAR-440): path resolvers and API-key precedence chain extracted
+// to `bootstrap::config`. Re-exported at this level so external paths
+// `crate::bootstrap::default_vault_path` and `crate::bootstrap::resolve_api_key`
+// stay valid (consumed by `admin::handlers`, `router`, `state`).
+pub(crate) use config::{default_vault_path, resolve_api_key};
+use config::default_allowlist_path;
 
 /// Build a fully-configured `AgentRuntime` from the application config.
 pub fn build_agent_runtime(config: &AppConfig) -> AgentRuntime {
@@ -2385,31 +2357,4 @@ mod tests {
         let _runtime = build_agent_runtime(&config);
     }
 
-    #[test]
-    fn resolve_api_key_prefers_config_over_env() {
-        // Config value should win when present
-        let result = resolve_api_key(
-            Some("from-config"),
-            "NONEXISTENT_VAULT_KEY",
-            "NONEXISTENT_ENV_VAR_12345",
-        );
-        assert_eq!(result, Some("from-config".to_string()));
-    }
-
-    #[test]
-    fn resolve_api_key_falls_back_to_env() {
-        // Set a unique env var for this test
-        let var_name = "GARRAIA_TEST_API_KEY_BOOTSTRAP_72";
-        // SAFETY: this test is single-threaded and uses a unique env var name.
-        unsafe { std::env::set_var(var_name, "from-env") };
-        let result = resolve_api_key(None, "NONEXISTENT_VAULT_KEY", var_name);
-        assert_eq!(result, Some("from-env".to_string()));
-        unsafe { std::env::remove_var(var_name) };
-    }
-
-    #[test]
-    fn resolve_api_key_returns_none_when_all_missing() {
-        let result = resolve_api_key(None, "NONEXISTENT_VAULT_KEY", "NONEXISTENT_ENV_VAR_99999");
-        assert_eq!(result, None);
-    }
 }
